requirements.txt
albumentations==1.3.0
matplotlib==3.6.1
numpy==1.23.4
omegaconf==2.2.1
opencv_python==4.6.0.66
pandas==1.5.1
Pillow==9.2.0
rich==12.6.0
streamlit==1.16.0
streamlit_drawable_canvas==0.9.2
tomli==2.0.1
torch==1.12.1
torchvision
# need to define some pytorch
https://download.pytorch.org/whl/cpu/torch-1.13.1%2Bcpu-cp310-cp310-linux_x86_64.whl
